VegaChart: add cleanup function in effect to avoid memory leak on component unmount
When the component unmount before Vega had the oportunity to render it leads to a an error due to the dom element not being present anymore.
Adding a cleanup function to check if the component is mounted enable to mitigate this issue. Also we avoid memory leaks by dereferencing vega listeners on unmount as part of this cleanup function

Refs: #214