stucture: add .stories to stories files
We add this ".stories" ending at the end of stories files because
with MDX, we might want to render custom component.

Because of the new way storybook CSF work, anything that is exported
in a file in /stories/* will be displayed in storybook.

We want a way to discriminate files in the stories/ folder that will not
be displayed by storybook.
That's why I change the `main.js` in order to filtrate this kind of files.
Now you will be able to create custom component to render in your MDX files.