stories: rename stories to *.stories.js
We add ".stories" ending at the end of stories files because
with MDX, we might want to render custom components that are not stories.

Currently, anything that is exported in a file in /stories/* will be
displayed in storybook.
We want a way to discriminate files in the stories/ folder that will not
be displayed by storybook.
So, I change the `main.js` (previous commit) in order to filtrate this kind of files.

Now you will be able to create custom component to render in your MDX files by creating
files that do not end with `.stories.js`.