move the aria label of select from input to the container
This commit fix a testing issue, it does not change anything about
the select component behavior.

react-select does not render our 'input' when the it's not 'searchable'.
-> the aria label definition is put on the container

In order to help for the testing when we have several selects, we want
to have an aria-label, it's define by the placeholder because it's
what the user see (vs the id)